requirements.txt
apache-airflow-providers-postgres
pandas
requests
sqlalchemy
psycopg2-binary
pyyaml